scripts/data_generator/requirements.txt
Faker==9.8.0
psycopg2==2.9.1
PyYAML==5.4.1
